Removing controllers
