+++ backend/src/routes/users/index.js
@@ -1,8 +1,28 @@
 const express = require('express')
-const controller = require('../../controllers/users')
+const ERRORS = require('../../errors')
+const User = require('../../models/User')
+const utils = require('../../utils')
 
 const router = express.Router()
 
-router.post('/', controller.register)
+router.post('/', register)
+function register(req, res, next) {
+    const { body: { username, password } } = req
+    if (!username || !password) {
+        next(ERRORS.MISSING_PARAMETERS(['username', 'password']))
+    } else {
+        User.register(username, password)
+            .then(user => ({
+                id: user.id,
+                username: user.username,
+                token: utils.genToken(user.id)
+            }))
+            .then(data => {
+                res.status(201)
+                res.json(data)
+            })
+            .catch(err => next(ERRORS.REGISTRATION_FAILED(err)))
+    }
+}
 
 module.exports = router
